# backend/app/api/static/docs/openapi-3.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Track, Manage, and Organize your Things.",
        "title": "Homebox API",
        "contact": {
            "name": "Homebox Team",
            "url": "https://discord.homebox.software"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/admin/settings": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Get site settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v1.AdminSettingsOut"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/admin/settings/algolia/reindex": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Trigger a full Algolia reindex",
                "responses": {
                    "202": {
                        "description": "Accepted"
                    }
                }
            }
        },
        "/v1/admin/settings/{section}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Update one site settings section",
                "parameters": [
                    {
                        "description": "Section name",
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "description": "Sparse section override",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v1.AdminSettingsOut"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Reset one site settings section to environment/default values",
                "parameters": [
                    {
                        "description": "Section name",
                        "name": "section",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v1.AdminSettingsOut"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "Bearer": {
                "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "config.AlgoliaConf": {
                "type": "object",
                "properties": {
                    "adminApiKey": {
                        "type": "string"
                    },
                    "appId": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "fields": {
                        "description": "Fields is a comma-separated allowlist of record fields to push. Empty\nmeans every field. objectID and groupId are always included.",
                        "type": "string"
                    },
                    "indexName": {
                        "type": "string"
                    },
                    "publicBaseUrl": {
                        "description": "PublicBaseURL is the externally reachable base URL used to build\nthumbnail links. Falls back to options.hostname when empty.",
                        "type": "string"
                    },
                    "publicImageUrls": {
                        "description": "PublicImageURLs includes an unauthenticated HMAC-signed thumbnail URL in\neach record so external search UIs can render item images.",
                        "type": "boolean"
                    },
                    "reindexInterval": {
                        "description": "ReindexInterval is a Go duration string (\"24h\"); kept as a string so it\nround-trips through the settings API and UI without nanosecond math.",
                        "type": "string"
                    }
                }
            },
            "config.BarcodeAPIConf": {
                "type": "object",
                "properties": {
                    "openFoodFactsContact": {
                        "type": "string"
                    },
                    "tokenBarcodespider": {
                        "type": "string"
                    }
                }
            },
            "config.LabelMakerConf": {
                "type": "object",
                "properties": {
                    "additionalInformation": {
                        "type": "string"
                    },
                    "boldFontPath": {
                        "type": "string"
                    },
                    "dynamicLength": {
                        "type": "boolean"
                    },
                    "fontSize": {
                        "type": "number"
                    },
                    "height": {
                        "type": "integer"
                    },
                    "labelServiceTimeout": {
                        "$ref": "#/components/schemas/time.Duration"
                    },
                    "labelServiceUrl": {
                        "type": "string"
                    },
                    "margin": {
                        "type": "integer"
                    },
                    "padding": {
                        "type": "integer"
                    },
                    "printCommand": {
                        "type": "string"
                    },
                    "regularFontPath": {
                        "type": "string"
                    },
                    "width": {
                        "type": "integer"
                    }
                }
            },
            "config.MailerConf": {
                "type": "object",
                "properties": {
                    "from": {
                        "type": "string"
                    },
                    "host": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "port": {
                        "type": "integer"
                    },
                    "username": {
                        "type": "string"
                    }
                }
            },
            "config.NotifierConf": {
                "type": "object",
                "properties": {
                    "allowNets": {
                        "description": "AllowNets will allow specific networks through for generic notifiers.\nIf this is filled, only these networks will be allowed through.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "blockBogonNets": {
                        "description": "BlockBogonNets will prevent generic notifiers from sending to bogon networks. (Reserved IPs)",
                        "type": "boolean"
                    },
                    "blockCloudMetadata": {
                        "description": "BlockCloudMetadata will prevent generic notifiers from sending to known cloud metadata IPs.",
                        "type": "boolean"
                    },
                    "blockLocalNets": {
                        "description": "BlockLocalNets will prevent generic notifiers from sending to local networks. (RFC1918)",
                        "type": "boolean"
                    },
                    "blockLocalhost": {
                        "description": "BlockLocalhost will prevent generic notifiers from sending to localhost.",
                        "type": "boolean"
                    },
                    "blockNets": {
                        "description": "BlockNets will block specific networks from generic notifiers.\nIf this is filled, these networks will be blocked from generic notifiers.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "config.Options": {
                "type": "object",
                "properties": {
                    "allowAnalytics": {
                        "type": "boolean"
                    },
                    "allowLocalLogin": {
                        "type": "boolean"
                    },
                    "allowRegistration": {
                        "type": "boolean"
                    },
                    "autoIncrementAssetId": {
                        "type": "boolean"
                    },
                    "currencyConfig": {
                        "type": "string"
                    },
                    "githubReleaseCheck": {
                        "type": "boolean"
                    },
                    "hostname": {
                        "type": "string"
                    },
                    "trustProxy": {
                        "type": "boolean"
                    }
                }
            },
            "config.Thumbnail": {
                "type": "object",
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    },
                    "height": {
                        "type": "integer"
                    },
                    "width": {
                        "type": "integer"
                    }
                }
            },
...
            "settings.Resolved": {
                "type": "object",
                "properties": {
                    "algolia": {
                        "$ref": "#/components/schemas/config.AlgoliaConf"
                    },
                    "barcode": {
                        "$ref": "#/components/schemas/config.BarcodeAPIConf"
                    },
                    "labelmaker": {
                        "$ref": "#/components/schemas/config.LabelMakerConf"
                    },
                    "mailer": {
                        "$ref": "#/components/schemas/config.MailerConf"
                    },
                    "notifier": {
                        "$ref": "#/components/schemas/config.NotifierConf"
                    },
                    "options": {
                        "$ref": "#/components/schemas/config.Options"
                    },
                    "thumbnail": {
                        "$ref": "#/components/schemas/config.Thumbnail"
                    }
                }
            },
...
            "time.Duration": {
                "type": "integer",
                "format": "int64",
                "enum": [
                    -9223372036854776000,
                    9223372036854776000,
                    1,
                    1000,
                    1000000,
                    1000000000,
                    60000000000,
                    3600000000000,
                    1,
                    1000,
                    1000000,
                    1000000000,
                    60000000000,
                    3600000000000
                ],
                "x-enum-varnames": [
                    "minDuration",
                    "maxDuration",
                    "Nanosecond",
                    "Microsecond",
                    "Millisecond",
                    "Second",
                    "Minute",
                    "Hour",
                    "Nanosecond",
                    "Microsecond",
                    "Millisecond",
                    "Second",
                    "Minute",
                    "Hour"
                ]
            },
...
            "v1.AdminSettingsOut": {
                "type": "object",
                "properties": {
                    "overridden": {
                        "description": "Overridden marks sections whose values come from the database rather\nthan environment variables / defaults.",
                        "type": "object",
                        "additionalProperties": {
                            "type": "boolean"
                        }
                    },
                    "settings": {
                        "$ref": "#/components/schemas/settings.Resolved"
                    }
                }
            },
...
        }
    }
}